refactor - test class 패키지 분리
