+++ src/test/java/racingcar/domain/CarListTest.java
@@ -1,9 +1,7 @@
-package racingcar;
+package racingcar.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.Car;
-import racingcar.domain.CarList;
 
 import java.util.ArrayList;
 import java.util.List;
