+++ src/test/java/racingcar/domain/GameNumbersTest.java
@@ -1,8 +1,7 @@
-package racingcar;
+package racingcar.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.GameNumbers;
 
 import static org.assertj.core.api.Assertions.assertThat;
